feat: View connected to model.
Map View updates with data from database.

+++ Pin+CoreDataClass.swift
@@ -11,11 +11,12 @@ import CoreData
 
 @objc(Pin)
 public class Pin: NSManagedObject {
-    convenience init(latitude: Float, longitude: Float, context: NSManagedObjectContext) {
+    convenience init(latitude: Double, longitude: Double, context: NSManagedObjectContext) {
         if let entity = NSEntityDescription.entity(forEntityName: "Pin", in: context) {
             self.init(entity: entity, insertInto: context)
             self.latitude = latitude
             self.longitude = longitude
+
         }else {
             print("Invalid entity for given context.")
             fatalError()    // Entity not accessible
